pyproject.toml
[project]
name = "django-roles-permissions"
version = "0.1.0"
description = "Django app with user roles, permissions, and groups"
authors = [
    {name = "mahimari555@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
